// enum/appointment.enum.ts
export enum AppointmentStatusEnum {
  SCHEDULED = 'scheduled',
  CONFIRMED = 'confirmed',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled',
  NO_SHOW = 'no-show',
}

export enum AppointmentTypeEnum {
  CONSULTATION = 'consultation',
  FOLLOW_UP = 'follow-up',
  LAB_TEST = 'lab-test',
  SURGERY = 'surgery',
  VACCINATION = 'vaccination',
}